fix some issues raised by rabbit
