+++ internal/cli/user_parse_local_file_nocgo.go
@@ -1,0 +1,25 @@
+//go:build !cgo
+
+//
+//  Copyright 2026 The InfiniFlow Authors. All Rights Reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+package cli
+
+import "fmt"
+
+func (c *CLI) UserParseLocalFile(cmd *Command) (ResponseIf, error) {
+	return nil, fmt.Errorf("parse local file is unavailable in the static CLI build")
+}
